Change Project Structure
Move src/api & data into a new module that allows for new apis to be wrapped.

* New mod.rs allowing export of cv19api
* Update README.md as examples have been renamed
* Move http request to utils.rs
