+++ examples/cv19api_deaths.rs
@@ -1,7 +1,7 @@
-use chrono::NaiveDate;
+use covid19::cv19api;
 
 fn main() {
-    let deaths = covid19::api::deaths_by_region();
+    let deaths = cv19api::api::deaths();
 
     if let Ok(resp) = deaths {
         println!("{:?}", resp.data[0]);
@@ -10,8 +10,8 @@ fn main() {
 
     let from = NaiveDate::from_ymd(2020, 4, 15);
     let to = NaiveDate::from_ymd(2020, 4, 25);
-    let use_ground_truth = true;
-    let parameterised = covid19::api::parameterised_deaths_by_region(from, to, use_ground_truth);
+    let use_ground_truth = false;
+    let parameterised = covid19::api::parameterised_deaths(from, to, use_ground_truth);
 
     println!("{:?}", parameterised)
 }
