+++ src/cv19api/mod.rs
@@ -1,0 +1,2 @@
+pub mod api;
+pub mod data;
